move sessions processor
